requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
pydantic
pydantic-settings
langchain
langchain-core
langchain-google-genai
langchain-groq
httpx
python-multipart
